@@ -7,14 +7,14 @@ homepage: https://github.com/gilmarsquinelato/FormKit_flutter
 documentation: https://github.com/gilmarsquinelato/FormKit_flutter
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  rxdart: ^0.25.0
+  rxdart: ^0.26.0-nullsafety.0
 
 dev_dependencies:
   flutter_test:
