@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:meta/meta.dart';
 import 'package:flutter/widgets.dart';
 import 'package:formkit/formkit.dart';
@@ -82,15 +83,15 @@ import 'package:formkit/formkit.dart';
 ///
 class FormKit extends StatefulWidget {
   const FormKit({
-    Key key,
-    @required this.child,
-    this.initialValues,
+    Key? key,
+    required this.child,
+    this.initialValues = const {},
     this.autovalidateMode = AutovalidateMode.disabled,
     this.onWillPop,
     this.validatorTimerMode,
     this.validatorInterval,
-  })  : assert(child != null),
-        super(key: key);
+    this.onSubmit,
+  }) : super(key: key);
 
   /// The widget below this widget in the tree.
   ///
@@ -100,7 +101,7 @@ class FormKit extends StatefulWidget {
   /// The initial values for each field under this form.
   ///
   /// Where the key is the field name with the reqpective field value as its value.
-  final Map<String, dynamic> initialValues;
+  final Map<String, dynamic?>? initialValues;
 
   /// Used to enable/disable form fields auto validation and update their error
   /// text.
@@ -112,7 +113,7 @@ class FormKit extends StatefulWidget {
   /// will make the fields use this value as default
   /// unless it's specified in the field properties
   /// overriding this value on the specific field.
-  final Duration validatorInterval;
+  final Duration? validatorInterval;
 
   /// {@macro formkit.fields.formKitField.validatorTimerMode}
   ///
@@ -120,7 +121,9 @@ class FormKit extends StatefulWidget {
   /// will make the fields use this value as default
   /// unless it's specified in the field properties
   /// overriding this value on the specific field.
-  final ValidatorTimerMode validatorTimerMode;
+  final ValidatorTimerMode? validatorTimerMode;
+
+  final Function(Map<String, dynamic?> values)? onSubmit;
 
   /// Enables the form to veto attempts by the user to dismiss the [ModalRoute]
   /// that contains the form.
@@ -132,11 +135,20 @@ class FormKit extends StatefulWidget {
   ///
   ///  * [WillPopScope], another widget that provides a way to intercept the
   ///    back button.
-  final WillPopCallback onWillPop;
+  final WillPopCallback? onWillPop;
 
   static FormKitState of(BuildContext context) {
     final scope = context.dependOnInheritedWidgetOfExactType<_FormKitScope>();
-    return scope._formState;
+    if (scope != null) {
+      return scope._formState;
+    }
+
+    throw FlutterError.fromParts(<DiagnosticsNode>[
+      ErrorSummary(
+          'FormKit.of() called with a context that does not contain a FormKit.'),
+      ErrorHint(
+          'The FormKit fields must be put as children of the FormKit widget'),
+    ]);
   }
 
   @override
@@ -147,16 +159,17 @@ class FormKitState extends State<FormKit> {
   int _generation = 0;
 
   final _fields = Set<FormKitFieldState>();
-  final _values = Map<String, dynamic>();
-  final _errors = Map<String, String>();
+  final _values = Map<String, dynamic?>();
+  final _errors = Map<String, String?>();
   final _fieldDependencies = Map<FormKitFieldState, Set<String>>();
+  final _dirtyFields = Set<FormKitFieldState>();
 
   @override
   void initState() {
     super.initState();
 
     if (widget.initialValues != null) {
-      _values.addAll(widget.initialValues);
+      _values.addAll(widget.initialValues!);
     }
   }
 
@@ -183,22 +196,45 @@ class FormKitState extends State<FormKit> {
         oldWidget.validatorInterval != widget.validatorInterval) {
       _fields.forEach((field) => field.createValidatorStreams());
     }
+
+    _updateFromInitialValues(oldWidget);
+  }
+
+  /// {@template formkit.formKit.submit}
+  /// This method calls the [validate] function,
+  /// if any error was found, it returns null,
+  /// otherwise it will call the [FormKit.onSubmit] callback
+  /// and return the current form values.
+  /// {@endtemplate}
+  Future<Map<String, dynamic>?> submit() async {
+    await validate();
+    if (hasErrors) {
+      return null;
+    }
+
+    if (widget.onSubmit != null) {
+      widget.onSubmit!(values);
+    }
+
+    return values;
   }
 
   /// {@template formkit.formKit.values}
   /// Returns the current values of the form
   /// {@endtemplate}
-  Map<String, dynamic> get values => {..._values};
+  Map<String, dynamic?> get values => {..._values};
 
   /// {@template formkit.formKit.setValues}
   /// Set the given values of the form fields based on their names
   /// {@endtemplate}
-  void setValues(Map<String, dynamic> values) {
+  void setValues(Map<String, dynamic?> values) {
     _values.addAll(values);
 
     for (final field in _fields) {
       field.setValue(values[field.name]);
     }
+
+    _forceUpdate();
   }
 
   /// {@template formkit.formKit.hasErrors}
@@ -210,7 +246,7 @@ class FormKitState extends State<FormKit> {
   /// {@template formkit.formKit.errors}
   /// The fields errors
   /// {@endtemplate}
-  Map<String, String> get errors => {..._errors};
+  Map<String, String?> get errors => {..._errors};
 
   /// {@template formkit.formKit.enqueueValidation}
   /// Enqueue the validation for all registered fields
@@ -228,27 +264,13 @@ class FormKitState extends State<FormKit> {
   /// It returns the validation result,
   /// if the value is [null] it indicates the field is valid
   /// {@endtemplate}
-  Future<Map<String, String>> validate() async {
+  Future<Map<String, String?>> validate() async {
     _errors.clear();
     _errors.addAll(await _validateFields(_fields));
 
-    setState(() {
-      _generation++;
-    });
+    _forceUpdate();
 
     return _errors;
-  }
-
-  Future<Map<String, String>> _validateFields(
-      Set<FormKitFieldState> fields) async {
-    final errors = Map<String, String>();
-
-    for (final field in fields) {
-      final name = field.name;
-      errors[name] = await field.validate(_values[name], _values);
-    }
-
-    return errors;
   }
 
   @internal
@@ -269,16 +291,31 @@ class FormKitState extends State<FormKit> {
   void unregister(FormKitFieldState field) => _fields.remove(field);
 
   @internal
-  void onFieldChanged(String name, dynamic value) {
-    _values[name] = value;
+  void onFieldChanged(FormKitFieldState field, dynamic value) {
+    _values[field.name] = value;
+    _dirtyFields.add(field);
 
     if (widget.autovalidateMode == AutovalidateMode.onUserInteraction) {
-      _validateDependentFields(name);
+      _validateDependentFields(field.name);
     }
+
+    _forceUpdate();
   }
 
   @internal
-  void onFieldValidated(String name, String error) => _errors[name] = error;
+  void onFieldValidated(String name, String? error) => _errors[name] = error;
+
+  Future<Map<String, String?>> _validateFields(
+      Set<FormKitFieldState> fields) async {
+    final errors = Map<String, String?>();
+
+    for (final field in fields) {
+      final name = field.name;
+      errors[name] = await field.validate(_values[name], _values);
+    }
+
+    return errors;
+  }
 
   void _validateDependentFields(String name) {
     _fieldDependencies.entries
@@ -289,20 +326,44 @@ class FormKitState extends State<FormKit> {
         // Enqueue the validation on each field
         .forEach((field) => field.enqueueValidation(_values[field.name]));
   }
+
+  void _updateFromInitialValues(FormKit oldWidget) {
+    if (mapEquals(oldWidget.initialValues, widget.initialValues)) {
+      return;
+    }
+
+    final newValues = widget.initialValues ?? {};
+    final unmodifiedFields =
+        _fields.where((field) => !_dirtyFields.contains(field));
+
+    for (final field in unmodifiedFields) {
+      _values[field.name] = newValues[field.name];
+      field.setValue(_values[field.name]);
+    }
+  }
+
+  void _forceUpdate() {
+    setState(() {
+      _generation++;
+    });
+  }
 }
 
 class _FormKitScope extends InheritedWidget {
   _FormKitScope({
-    Key key,
-    @required Widget child,
-    @required FormKitState formState,
-    @required int generation,
-  })  : _formState = formState,
+    Key? key,
+    required Widget child,
+    required FormKitState formState,
+    required int generation,
+  })   : _formState = formState,
         _generation = generation,
         super(key: key, child: child);
 
   final FormKitState _formState;
   final int _generation;
+
+  /// {@macro formkit.formKit.submit}
+  Future<Map<String, dynamic?>?> submit() => _formState.submit();
 
   /// {@macro formkit.formKit.values}
   Map<String, dynamic> get values => _formState.values;
@@ -311,13 +372,13 @@ class _FormKitScope extends InheritedWidget {
   void setValues(Map<String, dynamic> values) => _formState.setValues(values);
 
   /// {@macro formkit.formKit.validate}
-  Future<Map<String, String>> validate() => _formState.validate();
+  Future<Map<String, String?>> validate() => _formState.validate();
 
   /// {@macro formkit.formKit.enqueueValidation}
   void enqueueValidation() => _formState.enqueueValidation();
 
   /// {@macro formkit.formKit.errors}
-  Map<String, String> get errors => _formState.errors;
+  Map<String, String?> get errors => _formState.errors;
 
   /// {@macro formkit.formKit.hasErrors}
   bool get hasErrors => _formState.hasErrors;
